+++ src/main/java/com/jpage4500/devicemanager/manager/DeviceManager.java
@@ -993,7 +993,7 @@ public class DeviceManager implements RemoteConnectionManager.RemoteConnectionLi
             // convert list to serials
             List<String> serialList = new ArrayList<>();
             for (Device device : deviceList) serialList.add(device.serial);
-            progressListener.onProgress(0, 0, "Uploading " + Utils.bytesToDisplayString(file.length()) + "...");
+            progressListener.onProgress(0, 0, "Uploading " + FileUtils.bytesToDisplayString(file.length()) + "...");
             Result result = connection.installApp(serialList, file);
             if (listener != null) listener.onTaskComplete(result.isSuccess, result.result);
         });
@@ -1005,13 +1005,13 @@ public class DeviceManager implements RemoteConnectionManager.RemoteConnectionLi
 
     public Result installAppInternal(Device device, File file, ProgressListener progressListener) {
         if (device.remoteConnection != null) {
-            progressListener.onProgress(0, 0, "Uploading " + Utils.bytesToDisplayString(file.length()) + "...");
+            progressListener.onProgress(0, 0, "Uploading " + FileUtils.bytesToDisplayString(file.length()) + "...");
             return device.remoteConnection.installApp(List.of(device.serial), file);
         }
         Timer timer = new Timer();
-        log.trace("installAppInternal: file:{}, size:{}", file.getName(), Utils.bytesToDisplayString(file.length()));
+        log.trace("installAppInternal: file:{}, size:{}", file.getName(), FileUtils.bytesToDisplayString(file.length()));
         try {
-            if (progressListener != null) progressListener.onProgress(1, 4, "Uploading " + Utils.bytesToDisplayString(file.length()) + "...");
+            if (progressListener != null) progressListener.onProgress(1, 4, "Uploading " + FileUtils.bytesToDisplayString(file.length()) + "...");
             PackageManager packageManager = new PackageManager(device.jadbDevice);
             packageManager.install(file, (currentStep, totalSteps, message) -> {
                 // ignore step 1 "Uploading" since we're doing it above and with the file size
