+++ src/main/java/com/jpage4500/devicemanager/utils/FileUtils.java
@@ -5,9 +5,7 @@ import org.slf4j.LoggerFactory;
 
 import java.io.*;
 import java.text.DecimalFormat;
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
 
 public class FileUtils {
@@ -67,9 +65,8 @@ public class FileUtils {
     /**
      * return string description of number of bytes (45k, 320b, 1.1M)
      */
-    public static String bytesToDisplayString(Long sizeInBytes) {
-        if (sizeInBytes == null) return "";
-        else if (sizeInBytes <= 0) return String.valueOf(sizeInBytes);
+    public static String bytesToDisplayString(long sizeInBytes) {
+        if (sizeInBytes <= 0) return String.valueOf(sizeInBytes);
         int digitGroups = (int) (Math.log10(sizeInBytes) / Math.log10(1024));
         return sizeDisplayFormat.format(sizeInBytes / Math.pow(1024, digitGroups)) + SIZE_UNITS[digitGroups];
     }
@@ -118,6 +115,24 @@ public class FileUtils {
             boolean isOk = folder.delete();
             log.trace("deleteFolder: FILE: {} -> {}", isOk, folder.getAbsolutePath());
         }
+    }
+
+    /**
+     * Determine MIME type from filename
+     */
+    public static String getMimeType(String filename) {
+        if (filename == null) return null;
+        String lower = filename.toLowerCase();
+        if (lower.endsWith(".jpg") || lower.endsWith(".jpeg")) return "image/jpeg";
+        if (lower.endsWith(".png")) return "image/png";
+        if (lower.endsWith(".gif")) return "image/gif";
+        if (lower.endsWith(".txt")) return "text/plain";
+        if (lower.endsWith(".json")) return "application/json";
+        if (lower.endsWith(".xml")) return "application/xml";
+        if (lower.endsWith(".pdf")) return "application/pdf";
+        if (lower.endsWith(".zip")) return "application/zip";
+        if (lower.endsWith(".apk")) return "application/vnd.android.package-archive";
+        return "application/octet-stream";
     }
 
     public static class FileStats {
