--- src/main/java/com/jpage4500/devicemanager/utils/ReverseIterator.java
@@ -1,39 +1,0 @@
-package com.jpage4500.devicemanager.utils;
-
-import java.util.Iterator;
-import java.util.List;
-
-/**
- * iterate over a list from bottom to top
- */
-public class ReverseIterator<T> implements Iterator<T>, Iterable<T> {
-
-    private final List<T> list;
-    private int position;
-
-    public ReverseIterator(List<T> list) {
-        this.list = list;
-        this.position = list.size() - 1;
-    }
-
-    @Override
-    public Iterator<T> iterator() {
-        return this;
-    }
-
-    @Override
-    public boolean hasNext() {
-        return position >= 0;
-    }
-
-    @Override
-    public T next() {
-        return list.get(position--);
-    }
-
-    @Override
-    public void remove() {
-        throw new UnsupportedOperationException();
-    }
-
-}
